tap_pl: allocation of a tap using vsys
Planetlab tap interfaces are allocated by a vsys script (fd_tuntap).  The
script chooses the name of the interface and sends the corresponding fd
to user context through a Unix domain socket. The allocated tap interface
is not persistent. Moreover, the tap interface is not visibile in user
context until an IP address is assigned to it (by using vsys/vif_up).

The create_bridge script first creates a Planetlab tap with an IP
address, then creates a bridge with the same name of the tap.