list files that do not go into the distribution
The Makefile of Open vSwitch checks that all files in the
git repository are also distributed and stops with an error
otherwise. But we have some files in git that are not meant
to be distributed (i.e., installed on a sliver). These files can now
be listed in .non-distfiles and the new Makefile will ignore them.