self-contained tap creation in the switchd
This patch incorporates the functionality of pltap-ovs,
vsysc and create_bridge into the vswitch server itself.
Now the server is able to re-recreate any bridge contained
in the database after a restart.
The database must also contain the IP/NETMASK of the
tap device. These are stored as options of the bridge
internal interface.

Now the bridge name is no longer required to be the same
as the tap name. Instead, the tap interface remembers
the "real name" internally. The real name can be retrieved
with

ovs-appctl netdev-pltap/get-tapname

The sliver-ovs utility and the exp-tool/Makefile have
been updated accordingly.

TODO:
check wether we are allwed to:
- read the mac address of the tap;
- read the stats.
Otherwise see if we can do that via Vsys.