better handling of dependencies
Now bridges and interfaces survive a shutdown of the daemons.
Only port numbers are re-exchanged as needed on daemon restart.

This also fixes the 'make -j shutdown' bug.

Now, L/n1-n2 should mean that a link between node n1 and n2 has
been established in the db *and* it is also up (i.e., the daemons are
running).

TODO:
- there is some redundancy between start/stop and {del-,}{db,switch}s